Separate rsc by client and without-client
