+++ demo/client/dune
@@ -25,12 +25,12 @@
     --bundle
     --platform=browser
     --log-level=error
-    --outfile=app/demo/client/bundle.js))))
+    --outfile=app/demo/client/app.js))))
 
 (rule
  (alias client)
  (deps
-  (:file runtime.jsx)
+  (:file runtime-without-client.jsx)
   (alias_rec melange-app))
  (action
   (progn
@@ -40,4 +40,19 @@
     --bundle
     --platform=browser
     --log-level=error
-    --outfile=app/demo/client/rsc.js))))
+    --outfile=app/demo/client/rsc-without-client.js))))
+
+(rule
+ (alias client)
+ (deps
+  (:file runtime-with-client.jsx)
+  (alias_rec melange-app))
+ (action
+  (progn
+   (run
+    esbuild
+    %{file}
+    --bundle
+    --platform=browser
+    --log-level=error
+    --outfile=app/demo/client/rsc-with-client.js))))
