+++ demo/client/index.re
@@ -4,5 +4,5 @@ switch (ReactDOM.querySelector("#root")) {
 | Some(el) =>
   let _ = ReactDOM.Client.hydrateRoot(el, <App />);
   ();
-| None => ()
+| None => Js.log("No root element found")
 };
